requirements.txt
python-dotenv==1.0.0
agno==0.1.0
openai==1.12.0
exa-py==1.0.0
pydantic==2.5.2
rich==13.7.0
requests==2.31.0
googlemaps==4.10.0 